FLASK/requirement.txt
youtube-transcript-api==0.6.2
google-ai-generativelanguage==0.4.0
pathlib==1.0.1
python-dotenv==1.0.0
anaconda==0.0.1.1
